compress: field.py + organism.py → creature.py, real embeddings
The encounter and its memory were two files. They are one thing.

creature.py (1672 lines) absorbs both. The Field is the creature
sensing within a breath. The Organism is the creature persisting
across breaths. They share geometry, the FM client, the embedding
function. Making them one file says what was always true.

Real embeddings: default_embed_fn now tries to load
sentence-transformers via spark/local_embedder.py (all-MiniLM-L6-v2,
384-dim, CPU-only, already in this repo). Falls back to hash-seeded
random vectors when unavailable. The geometry becomes semantic or
stays decorative depending on which path runs.

On the Spark where local_embedder.py runs, the swap is automatic.
No config change needed. The creature opens its eyes.

field.py and organism.py become thin re-export shims.
All backward-compat import paths verified working.

4 files with real content:
  creature.py  1672  (the encounter and its memory)
  task_agent.py 573  (the thing that encounters)
  evolve.py     524  (selection across generations)
  run.py        495  (the shell)

Original 4106 → 3369 lines (-18%)
